Add Moving Spheres
