+++ Ray_trace_1/objects.h
@@ -20,8 +20,8 @@ int const MAX_THREADS = 3;
 #endif
 
 //  number of pixels, for each (axe x and y)
-int const WIDTH = 1000;
-int const HEIGHT = 600;
+int const WIDTH = 400;
+int const HEIGHT = 400;
 
 // Window size parameters
 int n1 = WIDTH;         
@@ -134,17 +134,17 @@ MATR operator * (MATR m1, MATR m2) {
     return m3;
 }
 
-#ifdef OMP
-double operator * (VEC3 v1, VEC3 v2) {
-    double answer = 0;
-#pragma omp parallel for reduction(+:answer)
-        for (int i = 0; i < 3; i++) {
-            answer += v1[i] * v2[i];
-        }
-#pragma omp barrier
-    return answer;
-}
-#else
+//#ifdef OMP
+//double operator * (VEC3 v1, VEC3 v2) {
+//    double answer = 0;
+//#pragma omp parallel for reduction(+:answer)
+//        for (int i = 0; i < 3; i++) {
+//            answer += v1[i] * v2[i];
+//        }
+//#pragma omp barrier
+//    return answer;
+//}
+//#else
 
 double operator * (VEC3 v1, VEC3 v2) {
 #ifdef GRAM
@@ -154,7 +154,7 @@ double operator * (VEC3 v1, VEC3 v2) {
     return answer;
 }
 
-#endif
+//#endif
 
 VEC3 operator * (double v1, VEC3 v2) {
     VEC3 V = { v2.x * v1, v2.y * v1, v2.z * v1 };
@@ -523,14 +523,13 @@ public:
     VEC3 get_center() {
         return center;
     }
+    void set_center(VEC3 p) {
+        center = p;
+    }
     virtual VEC3 get_norm(VEC3 P) override {
         VEC3 N = P - center;
         return N;
     }
-};
-
-class MovingSpheres : public SphereObj {
-    // TODO : realization
 };
 
 /// <summary>
@@ -881,308 +880,5 @@ public:
         std::swap(this->reflective, t.reflective);
         std::swap(this->specular, t.specular);
         return *this;
-    }
-};
-
-class Render final {
-    std::vector<GenericObject*>spheres;
-    std::vector<LightObj> lights;
-public:
-    // render camera motion parameters
-    double Pi = 3.1415926536;
-    VEC3 O = { 0, 1, 0 };                 // Ёто в общем точка, из которой лучи испускаютс€
-    VEC3 DIR = normalize({ 0, 0, 1 });  // Looking direction
-    double step = 0.5;                   // how much you will move // for debug, all logic must be rewrite
-    double angle = 0.005;                // how much you rotate
-
-    // winAPI variables
-    HINSTANCE hInst;                                // текущий экземпл€р
-    WCHAR szTitle[MAX_LOADSTRING];                  // “екст строки заголовка
-    WCHAR szWindowClass[MAX_LOADSTRING];            // им€ класса главного окна
-
-    //~Render() { for (auto i : spheres) { delete i; } } // FIXIT
-    Render(): hInst(nullptr) {};
-    Render(HINSTANCE hInstance) {
-        hInst = hInstance;
-        LoadStringW(hInstance, IDS_APP_TITLE, szTitle, MAX_LOADSTRING);
-        LoadStringW(hInstance, IDC_RAYTRACE1, szWindowClass, MAX_LOADSTRING);
-    };
-
-    /// <summary>
-    /// Use this function to set objects, that will be in render
-    /// If you send object to render, it will now manage memory
-    /// </summary>
-    /// <param name="id"></param>
-    /// <returns></returns>
-    /// 
-    void set_obj(GenericObject* object) {
-        spheres.push_back(object);
-    }
-
-    void set_light(LightObj* object) {
-        lights.push_back(*object);
-    }
-
-    void set_position(VEC3 v) {
-        O = v;
-    }
-
-    void set_direction(VEC3 v) {
-        DIR = normalize(v);
-    }
-    
-    GenericObject* get_object(int id) { // FIXIT (pointless)
-        if (spheres.size() > id) {
-            return spheres[id];
-        }
-        else {
-            return nullptr;
-        }
-    }
-
-    void delete_spheres(){
-        for (auto i : spheres) {
-            delete i;
-        }
-        spheres.clear();
-    }
-   
-    ///
-    /// Register window class
-    /// 
-    ATOM MyRegisterClass(LRESULT CALLBACK WndProc(HWND, UINT, WPARAM, LPARAM))
-    {
-        WNDCLASSEXW wcex;
-
-        wcex.cbSize = sizeof(WNDCLASSEX);
-
-        wcex.style = CS_HREDRAW | CS_VREDRAW;
-        wcex.lpfnWndProc = WndProc;
-        wcex.cbClsExtra = 0;
-        wcex.cbWndExtra = 0;
-        wcex.hInstance = hInst;
-        wcex.hIcon = LoadIcon(hInst, MAKEINTRESOURCE(IDI_RAYTRACE1));
-        wcex.hCursor = LoadCursor(nullptr, IDC_ARROW);
-        wcex.hbrBackground = (HBRUSH)(COLOR_WINDOW + 1);
-        wcex.lpszMenuName = MAKEINTRESOURCEW(IDC_RAYTRACE1);
-        wcex.lpszClassName = szWindowClass;
-        wcex.hIconSm = LoadIcon(wcex.hInstance, MAKEINTRESOURCE(IDI_SMALL));
-
-        return RegisterClassExW(&wcex);
-    }
-
-//
-//   ‘”Ќ ÷»я: InitInstance(HINSTANCE, int)
-//
-//   ÷≈Ћ№: —охран€ет маркер экземпл€ра и создает главное окно
-//
-//    ќћћ≈Ќ“ј–»»:
-//
-//        ¬ этой функции маркер экземпл€ра сохран€етс€ в глобальной переменной, а также
-//        создаетс€ и выводитс€ главное окно программы.
-//
-    BOOL InitInstance(int nCmdShow)
-    {
-        HWND hWnd = CreateWindowW(szWindowClass, szTitle, WS_OVERLAPPEDWINDOW,
-            CW_USEDEFAULT, 0, n1, n2, nullptr, nullptr, hInst, nullptr);
-        ShowWindow(hWnd, nCmdShow);
-        UpdateWindow(hWnd);
-        return TRUE;
-    }
-
-    void DrawScene(HWND hWnd) {
-        PAINTSTRUCT ps;
-        HDC hdc = BeginPaint(hWnd, &ps);
-
-        RECT rect;
-        GetClientRect(hWnd, &rect);
-
-        //pixel = rect.right / WIDTH;
-        //n1 = rect.right - rect.right % pixel + pixel;
-        //n2 = rect.bottom - rect.bottom % pixel + pixel;
-        //double h = (double)n2 / n1 * d;
-
-
-        HDC hmdc = CreateCompatibleDC(hdc);
-        HBITMAP bit = CreateCompatibleBitmap(hdc, n1, n2);
-        SelectObject(hmdc, bit);
-
-        /*
-        BITMAPINFO bif;
-        ZeroMemory(&bif, sizeof(BITMAPINFO));
-        bif.bmiHeader.biSize = sizeof(BITMAPINFOHEADER);
-        bif.bmiHeader.biBitCount = 32;
-        bif.bmiHeader.biWidth = n1;
-        bif.bmiHeader.biHeight = n2;
-        bif.bmiHeader.biPlanes = 1;
-        bif.bmiHeader.biCompression = BI_RGB;
-        bif.bmiHeader.biClrImportant = 0;
-        */
-#ifdef clock_1
-        clock_t start = clock();
-#endif
-        COL c; VEC3 D;
-        int i; int j;
-        RECT l;
-
-        for (int k = 0; k < n1 * n2; k += pixel) {
-            i = k / n2 * pixel;
-            j = k % n2;
-            D = -cross(DIR, UP) * ((i - (double)n1 / 2) * w / (double)n1) - UP * ((j - (double)n2 / 2) * h / (double)n2) + DIR * d;
-            c = Trace(O, D, 0.01, positive_inf, depth);  // set_color
-            //HBRUSH hb = CreateSolidBrush(RGB(c[0], c[1], c[2]));
-            l.left = i; l.top = j; l.right = i + pixel;  l.bottom = j + pixel;
-            PaintRect(hmdc, &l, RGB(c[0], c[1], c[2]));
-            //FillRect(hmdc, &l, hb); // this is huge
-            //DeleteObject(hb);
-            //SetPixel(hmdc, i, j, RGB(c[0], c[1], c[2]));
-        }
-
-        BitBlt(hdc, 0, 0, n1, n2, hmdc, 0, 0, SRCCOPY); // fast, 0ms
-
-        //GetDIBits(hmdc, bit, 0, 0, 0, &bif, DIB_RGB_COLORS);
-        //GetDIBits(hmdc, bit, 0, n2, im, &bif, DIB_RGB_COLORS);
-        //SetDIBitsToDevice(hdc, 0, 0, n1, n2, 0, 0, 0, n2, im, &bif, DIB_RGB_COLORS);
-        //SelectObject(hmdc, bit);
-
-        //BitBlt(hdc, 0, 0, n1, n2, hmdc, 0, 0, SRCCOPY);
-#ifdef clock_1
-        clock_t end = clock();
-        clock_t result = end - start;
-#endif
-        DeleteObject(bit);
-        DeleteDC(hmdc);
-
-        EndPaint(hWnd, &ps);
-        UpdateWindow(hWnd);
-    }
-
-    void CameraRotate(HWND &hWnd) 
-    {
-        POINT p;
-        RECT rectangle;
-        LPPOINT point = &p;
-        LPRECT rect = &rectangle;
-        GetCursorPos(point);
-        GetWindowRect(hWnd, rect);
-        int delta_y = point->x - (rect->right + rect->left) / 2;
-        int delta_x = point->y - (rect->bottom + rect->top) / 2;
-        if (abs(DIR.x) < abs(DIR.z)) {
-            if (DIR.z > 0) {
-                DIR = normalize(xRotate(delta_x * angle, DIR));
-            }
-            else {
-                DIR = normalize(xRotate(-delta_x * angle, DIR));
-            }
-        }
-        else {
-            if (DIR.x > 0) {
-                DIR = normalize(zRotate(-delta_x * angle, DIR));
-            }
-            else {
-                DIR = normalize(zRotate(delta_x * angle, DIR));
-            }
-        }
-        DIR = normalize(yRotate(delta_y * angle, DIR));
-
-
-        SetCursorPos((rect->right + rect->left) / 2, (rect->bottom + rect->top) / 2);
-    }
-
-    void ClosestIntersection(VEC3 O, VEC3 V, double t_min, double t_max, GenericObject*& closest_sphere, double& closest_t) {
-        double t;
-        GenericObject* Sphere;
-        for (int j = 0; j < spheres.size(); j++) {
-            Sphere = spheres[j];
-            Sphere->Intercections(O, V, t_min, t_max, t);  
-            if (((t >= t_min) && (t <= t_max)) && (t < closest_t)) {
-                closest_t = t;
-                closest_sphere = (spheres[j]);
-            }
-        }
-    }
-
-    /// <summary>
-    /// </summary>
-    /// <param name="P - the intersection point of the ray and the object"></param>
-    /// <param name="N - normal for the surface in the point P"></param>
-    /// <param name="V - vector from the P to the source of ray (point O)"></param>
-    /// <param name="s - specularity of the object (матовость)"></param>
-    /// <returns></returns>
-    double ComputeLight(VEC3& P, VEC3& N, VEC3 V, double s) {
-        // intensity of the initial color, after computing light
-        double i = 0.0;
-
-        // lights - vector<LightObj>
-        for (LightObj &Light: lights) {
-            if (Light.get_type() == 'a') {
-                i += Light.get_intensity();
-            }
-            else {
-                VEC3 L;                                                                     // vector to the light
-                if (Light.get_type() == 'p') {                                             // FIXIT (refactor Light to be inheritance
-                    L = Light.get_d() - P;
-                }
-                if (Light.get_type() == 'd') {
-                    L = Light.get_d();
-                }
-
-                double closest_t = positive_inf;                                           // parameter of the distamce to the shadow_sphere
-                GenericObject* shadow_sphere = nullptr;                                    // shadow_sphere - object, which will make a shadow
-                ClosestIntersection(P, L, ray_epsilon, positive_inf, shadow_sphere, closest_t);// calculate intersection
-                
-                if (shadow_sphere != nullptr) {                                            // if the shadow_sphere != nullptr, then there is a shadow
-                    continue;
-                }
-                // if not, then compute light
-                double n_dot_l = N * L;
-                if (n_dot_l > 0) {
-                    if ((N*N != 0) && (L*L != 0)) {
-                        i += Light.get_intensity() * n_dot_l / abs(N) / abs(L);
-                    }
-                }
-
-                if (s != -1) {
-                    VEC3 R = ReflectRay(L, N);
-                    double r_dot_v = (R * V);
-                    if ((r_dot_v > 0) && (R*R != 0) && (V*V != 0)) {
-                        i += Light.get_intensity() * std::pow(r_dot_v / (abs(R) * abs(V)), s);// this is expensive FIXIT
-                    }
-                }
-            }
-
-        }
-        return i < 1 ? i : 1;
-    }
-
-    COL Trace(VEC3 O, VEC3 V, double t_min, double t_max, int depth) {
-        double closest_t = positive_inf;
-        GenericObject* closest_sphere = nullptr;
-
-        if (spheres.size() > 0) 
-        {
-            ClosestIntersection(O, V, t_min, t_max, closest_sphere, closest_t);
-        }
-
-        if (closest_sphere == nullptr) 
-        {
-            return BG_C;
-        }
-
-        VEC3 P = O + closest_t * V;
-        VEC3 N = closest_sphere->get_norm(P);
-
-        if (abs(N) != 0) { N = normalize(N); }
-
-        COL local_color = ComputeLight(P, N, -V, closest_sphere->get_specular()) * closest_sphere->get_color();
-
-        double r = closest_sphere->get_reflective();
-
-        if ((depth <= 0) || (r <= 0)) {
-            return local_color;
-        }
-
-        COL reflected_color = Trace(P, ReflectRay(-V, N), epsilon, positive_inf, depth - 1);
-        return (1 - r) * local_color + r * reflected_color;
     }
 };
